% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_surv.R
\name{as_surv}
\alias{as_surv}
\alias{as_surv.default}
\alias{as_surv.event_time}
\title{Convert to a survival object}
\usage{
as_surv(x, ...)

\method{as_surv}{default}(x, ...)

\method{as_surv}{event_time}(x, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Not currently used.}
}
\value{
A \code{Surv} object.
}
\description{
\code{as_surv()} converts an object to the format produced by
\code{\link[survival:Surv]{survival::Surv()}}.
}
\details{
For \code{event_time} vectors, the type of the result depends on the censoring
types present:
\itemize{
\item Only exact and right-censored values: \code{type = "right"}.
\item Only exact and left-censored values: \code{type = "left"}.
\item Any other combination: \code{type = "interval"}.
}

Missing values in \code{x} are missing in the result.
}
\examples{
\dontshow{if (rlang::is_installed("survival")) withAutoprint(\{ # examplesIf}
x <- event_time(
  time = c(7, 5, 3, 2),
  status = c("e", "r", "l", "i"),
  time_max = c(NA, NA, NA, 4)
)
as_surv(x)

as_surv(event_time(c(7, 5), c("e", "r")))

if (rlang::is_installed(c("dplyr", "survival"))) {
  library(dplyr)
  library(survival)

  set.seed(1)
  tibble(times = sort(rexp(5)), status = c("l", "e", "e", "e", "r")) |>
    mutate(
      event_time = event_time(times, status),
      surv_obj = as_surv(event_time)
    )
}
\dontshow{\}) # examplesIf}
}
