% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.event_time}
\alias{as_tibble.event_time}
\title{Convert an event time vector to a tibble}
\usage{
\method{as_tibble}{event_time}(x, ...)
}
\arguments{
\item{x}{An \code{event_time} vector.}

\item{...}{Not currently used.}
}
\value{
A tibble with columns \code{time}, \code{status}, and \code{time_max}. \code{time_max} is
always included and is missing for values that are not interval censored.
}
\description{
The reverse of \code{\link[=event_time]{event_time()}}: split an \code{event_time} vector into its \code{time},
\code{status}, and \code{time_max} components.
}
\examples{
\dontshow{if (rlang::is_installed("tibble")) withAutoprint(\{ # examplesIf}
x <- event_time(c(7, 5, 2), c("e", "r", "i"), c(NA, NA, 4))
tibble::as_tibble(x)
\dontshow{\}) # examplesIf}
}
