% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etd-package.R
\docType{package}
\name{etd-package}
\alias{etd}
\alias{etd-package}
\title{etd: Event Time Data}
\description{
Create vector types for potentially event times.
}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
