% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_time.R
\name{event_time}
\alias{event_time}
\title{Create a vector of event times}
\usage{
event_time(time, status, time_max = NULL, call = rlang::current_env())
}
\arguments{
\item{time}{A non-negative numeric vector (double or integer) of event times.
For interval-censored values, this is the lower end of the interval.}

\item{status}{A character vector the same length as \code{time}. Possible values
are \code{"e"} (complete/exact), \code{"r"} (right censored), \code{"l"} (left censored),
and \code{"i"} (interval censored). It can only be missing where \code{time} is
missing.}

\item{time_max}{A numeric vector the same length as \code{time} with the upper end
of the interval for interval-censored values. It must be missing for
values that are not interval censored, and greater than \code{time} otherwise.
The default \code{NULL} sets all values to missing.}

\item{call}{The execution environment of a currently running function, used
in error messages.}
}
\value{
An \code{event_time} vector. Each element of its \code{time} field is a single
number, or a sorted pair of numbers for interval-censored values.
}
\description{
\code{event_time()} creates a vector of event times that may be exact times, or
right/left/interval censored.
}
\examples{
x <- event_time(
  time = c(7, 5, 3, 2, NA),
  status = c("e", "r", "l", "i", NA),
  time_max = c(NA, NA, NA, 4, NA)
)
x

x[2:4]

if (rlang::is_installed(c("dplyr"))) {
  library(dplyr)

  set.seed(1)
  tibble(times = sort(rexp(5)), status = c("l", "e", "e", "e", "r")) |>
    mutate(event_time = event_time(times, status))
}
}
