% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_time}
\alias{extract_time}
\alias{extract_time.event_time}
\alias{extract_status}
\alias{extract_status.event_time}
\title{Extract components of an event time vector}
\usage{
extract_time(x, ...)

\method{extract_time}{event_time}(x, ...)

extract_status(x, ...)

\method{extract_status}{event_time}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}
}
\value{
\itemize{
\item \code{extract_time()}: a double vector when no values are interval censored.
Otherwise, a two-column matrix with columns \code{time} and \code{time_max}, where
\code{time_max} is missing for values that are not interval censored.
\item \code{extract_status()}: a character vector of status codes (\code{"e"}, \code{"r"},
\code{"l"}, or \code{"i"}).
}
}
\description{
\code{extract_time()} returns the event times and \code{extract_status()} returns the
censoring status codes.
}
\examples{
x <- event_time(c(7, 5, 2), c("e", "r", "i"), c(NA, NA, 4))
extract_time(x)
extract_status(x)

extract_time(x[1:2])
}
