% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_time.R
\name{new_event_time}
\alias{new_event_time}
\title{Low-level constructor for event time vectors}
\usage{
new_event_time(
  time = list(),
  status = character(),
  call = rlang::current_env()
)
}
\arguments{
\item{time}{A list. Each element is a single double, or a sorted pair of
doubles for interval-censored values.}

\item{status}{A character vector of status codes.}

\item{call}{The execution environment of a currently running function, used
in error messages.}
}
\value{
An \code{event_time} vector.
}
\description{
\code{new_event_time()} creates an \code{event_time} vector with only minimal type
checks. Use \code{\link[=event_time]{event_time()}} to create a validated vector.
}
\examples{
new_event_time(list(7, c(2, 4)), c("e", "i"))
}
