% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etd-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tibble}{\code{\link[tibble:as_tibble]{as_tibble()}}}
}}

